@@ -11,41 +11,37 @@
         "calebporzio/sushi": "^2.4",
         "doctrine/dbal": "^3.5",
         "guzzlehttp/guzzle": "^7.4",
-        "laravel/framework": "^10.1",
-        "laravel/jetstream": "^3.0",
-        "laravel/sanctum": "^3.2",
-        "laravel/tinker": "^2.8",
+        "laravel/framework": "^11.2",
+        "laravel/tinker": "^2.9",
         "league/flysystem-ftp": "^3.0",
-        "livewire/livewire": "^3.0",
-        "owenvoke/blade-fontawesome": "^2.2",
+        "livewire/livewire": "^3.3",
+        "owenvoke/blade-fontawesome": "^2.6",
         "spatie/data-transfer-object": "^3.7",
         "spatie/icalendar-generator": "^2.3",
-        "spatie/laravel-backup": "^8.1",
-        "spatie/laravel-feed": "^4.2",
-        "spatie/laravel-ignition": "^2.0",
+        "spatie/laravel-backup": "^8.6",
+        "spatie/laravel-feed": "^4.4",
+        "spatie/laravel-ignition": "^2.4",
         "spatie/laravel-login-link": "^1.2",
-        "spatie/laravel-mailcoach-mailer": "^1.0",
-        "symfony/http-client": "^6.2",
-        "symfony/postmark-mailer": "^6.2",
-        "vinkla/hashids": "^11.0"
+        "spatie/laravel-mailcoach-mailer": "^1.2",
+        "symfony/http-client": "^7.0",
+        "symfony/postmark-mailer": "^7.0",
+        "vinkla/hashids": "^12.0"
     },
     "require-dev": {
-        "brianium/paratest": "^6.4",
-        "christophrumpel/missing-livewire-assertions": "^2.0",
-        "fakerphp/faker": "^1.17",
+        "christophrumpel/missing-livewire-assertions": "^2.5",
+        "fakerphp/faker": "^1.23",
         "friendsofphp/php-cs-fixer": "^3.6",
-        "laravel/sail": "^1.18",
-        "mockery/mockery": "^1.5",
-        "nunomaduro/collision": "^6.1",
-        "nunomaduro/larastan": "^2.4",
+        "laravel/sail": "^1.26",
+        "mockery/mockery": "^1.6",
+        "nunomaduro/collision": "^8.0 ",
+        "nunomaduro/larastan": "^2.8",
         "nunomaduro/patrol": "^1.1",
-        "pestphp/pest": "^1.21",
-        "pestphp/pest-plugin-laravel": "^1.4",
-        "pestphp/pest-plugin-parallel": "^1.0",
+        "pestphp/pest": "^2.34",
+        "pestphp/pest-plugin-laravel": "^2.3",
+        "phpunit/phpunit": "^10.0",
         "roave/security-advisories": "dev-latest",
-        "spatie/laravel-ray": "^1.32",
-        "spatie/test-time": "^1.3",
-        "phpunit/phpunit": "^9.5.19"
+        "spatie/laravel-ray": "^1.35",
+        "spatie/test-time": "^1.3"
     },
     "config": {
         "optimize-autoloader": true,
@@ -73,7 +69,7 @@
             "Tests\\": "tests/"
         }
     },
-    "minimum-stability": "stable",
+    "minimum-stability": "dev",
     "prefer-stable": true,
     "scripts": {
         "post-autoload-dump": [
