@@ -4,8 +4,9 @@ use App\Models\Channel;
 use App\Models\Stream;
 use Carbon\Carbon;
 use Vinkla\Hashids\Facades\Hashids;
+use function Pest\Laravel\get;
 
-it('shows only finished streams', function() {
+it('shows only finished streams', function () {
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -26,13 +27,13 @@ it('shows only finished streams', function() {
         ->create();
 
     // Act & Assert
-    $this->get(route('archive'))
+    get(route('archive'))
         ->assertDontSee('Live stream')
         ->assertDontSee('Upcoming stream')
         ->assertSee('Finished stream');
 });
 
-it('uses actual start time first for order if given', function() {
+it('uses actual start time first for order if given', function () {
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -51,14 +52,14 @@ it('uses actual start time first for order if given', function() {
         ->create();
 
     // Act & Assert
-    $this->get(route('archive'))
+    get(route('archive'))
         ->assertSeeInOrder([
             'Finished latest',
             'Finished oldest',
         ]);
 });
 
-it('orders streams from latest to oldest', function() {
+it('orders streams from latest to oldest', function () {
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -82,7 +83,7 @@ it('orders streams from latest to oldest', function() {
         ->create();
 
     // Act & Assert
-    $this->get(route('archive'))
+    get(route('archive'))
         ->assertSeeInOrder([
             'Finished one day ago',
             'Finished two days ago',
@@ -90,7 +91,7 @@ it('orders streams from latest to oldest', function() {
         ]);
 });
 
-it('does not show deleted streams', function() {
+it('does not show deleted streams', function () {
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -111,7 +112,8 @@ it('does not show deleted streams', function() {
         ->assertDontSee('Stream deleted');
 });
 
-it('shows duration of stream if given', function() {
+it('shows duration of stream if given', function () {
+    $this->withoutExceptionHandling();
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -121,12 +123,11 @@ it('shows duration of stream if given', function() {
         ->create();
 
     // Act & Assert
-    $this
-        ->get(route('archive'))
+    get(route('archive'))
         ->assertSee('1h 12m');
 });
 
-it('searches for streams on title', function() {
+it('searches for streams on title', function () {
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -147,7 +148,8 @@ it('searches for streams on title', function() {
         ->create();
 
     // Act & Assert
-    $this->get(route('archive', ['search' => 'three']))
+    get(route('archive', ['search' => 'three']))
+        ->assertOk()
         ->assertSee([
             'Stream Three',
         ])->assertDontSee([
@@ -156,7 +158,7 @@ it('searches for streams on title', function() {
         ]);
 });
 
-it('searches for streams on channel title', function() {
+it('searches for streams on channel title', function () {
     // Arrange
     $channelShown = Channel::factory()->create(['name' => 'Channel Shown']);
     Stream::factory()
@@ -178,7 +180,8 @@ it('searches for streams on channel title', function() {
         ->create();
 
     // Act & Assert
-    $this->get(route('archive', ['search' => 'Channel Shown']))
+    get(route('archive', ['search' => 'Channel Shown']))
+        ->assertOk()
         ->assertSee([
             'Stream Shown #1',
             'Stream Shown #2',
@@ -192,7 +195,7 @@ it('searches for streams on channel title', function() {
 it('searches
  for streams by specific
  ->$this->withTitle()
-  streamer', function() {
+  streamer', function () {
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -207,12 +210,13 @@ it('searches
         ->create();
 
     // Act & Assert
-    $this->get(route('archive', ['streamer' => Hashids::encode(1)]))
+    get(route('archive', ['streamer' => Hashids::encode(1)]))
+        ->assertOk()
         ->assertSee('Stream Shown')
         ->assertDontSee('Stream Hidden');
 });
 
-it('searches for streams by specific streamer and search term', function() {
+it('searches for streams by specific streamer and search term', function () {
     // Arrange
     $channel = Channel::factory()->create();
     Stream::factory()
@@ -228,12 +232,13 @@ it('searches for streams by specific streamer and search term', function() {
         ->create();
 
     // Act & Assert
-    $this->get(route('archive', ['streamer' => '1', 'search' => 'Shown']))
+    get(route('archive', ['streamer' => '1', 'search' => 'Shown']))
+        ->assertOk()
         ->assertSee('Stream Shown')
         ->assertDontSee('Stream Hidden');
 });
 
-it('orders streamers by name', function() {
+it('orders streamers by name', function () {
     $this->withoutExceptionHandling();
     // Arrange
     Channel::factory()
@@ -250,7 +255,8 @@ it('orders streamers by name', function() {
         ->create(['name' => 'Caleb Porzio']);
 
     // Act & Assert
-    $this->get(route('archive'))
+    get(route('archive'))
+        ->assertOk()
         ->assertSeeInOrder([
             'Adrian Nürnberger',
             'Caleb Porzio',
