@@ -7,7 +7,7 @@ use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
 
 abstract class TestCase extends BaseTestCase
 {
-    use CreatesApplication, LazilyRefreshDatabase;
+    use LazilyRefreshDatabase;
 
     private string $originalYoutubeApiKey;
 
